ajout des autre DELETE

+++ api/users.php
@@ -4,14 +4,15 @@
 include '../includes/header.php';
 
 if ($_SERVER['REQUEST_METHOD'] == "GET") {
-		//if ($_GET['url'] == "users") {
+		
+
         	$requete = $pdo->prepare("SELECT * FROM users");
 			$requete->execute();
 			$resultats = $requete->fetchAll();
 			$retour["Utilisateur"]["donnee"] = $resultats; 
 			$retour["Utilisateur"]["nb d'utilisateur"] = count($resultats);
 
-        	retour_json(true,'Données utilisateur',$retour);
+        	retour_json(true,'Donnees utilisateur',$retour);
         	http_response_code(200);
         }
 
@@ -38,16 +39,19 @@ else if ($_SERVER['REQUEST_METHOD'] == "PUT") {}
 
 */
 else if ($_SERVER['REQUEST_METHOD'] == "DELETE") {
-	 //Tu recuperes l'id du contact
-         $id = $_GET['id'];
-         //Requete SQL pour supprimer le contact dans la base
-        
+	if(isset($_GET['id'])&& !empty( $_GET['id'] )){
+        $id=$_GET['id'];
+
+	 
+         //Requete SQL pour supprimer l'user dans la base
         $req = "DELETE FROM users WHERE users_id = '$id'";
+          
         //execution de la requete et obtention d'un statement
         $stmt = $pdo->query($req);
 
         //mode de fetch
         $stmt->setFetchMode(PDO::FETCH_OBJ);
+        http_response_code(200);        
 
         $requete = $pdo->prepare("SELECT * FROM users");
 			$requete->execute();
@@ -56,14 +60,13 @@ else if ($_SERVER['REQUEST_METHOD'] == "DELETE") {
 			$retour["Utilisateur"]["nb d'utilisateur"] = count($resultats);
 
         	retour_json(true,'Données utilisateur',$retour);
-        	http_response_code(200);
         }
 
     else {
-    	http_response_code(404);    }
+    	http_response_code(404);
+    	retour_json(false,'ID NON VALIDE');
+
+    }
+}
     
-
-         //Et la tu rediriges vers ta page contacts.php pour rafraichir la liste
-        // header("location:users.php");
-
 ?>
